Fix merge bugs
