+++ chains/evm/calls/utils.go
@@ -172,12 +172,3 @@ func Simulate(c SimulateCallerClient, block *big.Int, txHash common.Hash, from c
 	log.Debug().Msg(string(bs))
 	return bs, nil
 }
-
-func ConstructErc721DepositData(tokenId *big.Int, destRecipient []byte) []byte {
-	var data []byte
-	data = append(data, math.PaddedBigBytes(tokenId, 32)...)                               // Resource Id + Token Id
-	data = append(data, math.PaddedBigBytes(big.NewInt(int64(len(destRecipient))), 32)...) // Length of recipient
-	data = append(data, destRecipient...)                                                  // Recipient
-
-	return data
-}
